+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/swapxxx/EvmBlockchainHelper.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.modules.swap
+package io.horizontalsystems.bankwallet.modules.swapxxx
 
 import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.core.ethereum.EvmCoinServiceFactory
